requirements.txt
streamlit
arxiv
python-dotenv
anthropic